refactor: replace tabs with bottom drawer in dashboard metric card
